requirements.txt
# requirements.txt
python-telegram-bot[job-queue]>=20.0  # Use version 20+ for async/await syntax
aiohttp>=3.8.0
Flask>=2.0.0
python-dotenv>=0.20.0
requests # Keep for Flask example compatibility if needed, but aiohttp is preferred for async